backend/requirements.txt
flask==3.1.0
flask-cors==4.0.1
firecrawl-py==4.3.6
requests==2.32.5
python-dotenv==1.1.1
pathlib==1.0.1
